Add (disabled) TriggerRunClientLiveTests
