+++ sdk/synapse/Azure.Analytics.Synapse.Artifacts/tests/TriggerClientLiveTests.cs
@@ -21,37 +21,6 @@ namespace Azure.Analytics.Synapse.Artifacts.Tests
     /// </remarks>
     public class TriggerClientLiveTests : RecordedTestBase<SynapseTestEnvironment>
     {
-            internal class DisposableTrigger : IAsyncDisposable
-        {
-            private readonly TriggerClient _client;
-            public TriggerResource Resource;
-
-            private DisposableTrigger (TriggerClient client, TriggerResource resource)
-            {
-                _client = client;
-                Resource = resource;
-            }
-
-            public string Name => Resource.Name;
-
-            public static async ValueTask<DisposableTrigger> Create (TriggerClient client, TestRecording recording) =>
-                new DisposableTrigger (client, await CreateResource(client, recording));
-
-            public static async ValueTask<TriggerResource> CreateResource (TriggerClient client, TestRecording recording)
-            {
-                string triggerName = recording.GenerateId("Trigger", 16);
-                TriggerResource trigger = new TriggerResource(new ScheduleTrigger(new ScheduleTriggerRecurrence()));
-                TriggerCreateOrUpdateTriggerOperation createOperation = await client.StartCreateOrUpdateTriggerAsync (triggerName, trigger);
-                return await createOperation.WaitForCompletionAsync ();
-            }
-
-            public async ValueTask DisposeAsync()
-            {
-                TriggerDeleteTriggerOperation deleteOperation = await _client.StartDeleteTriggerAsync (Name);
-                await deleteOperation.WaitForCompletionAsync ();
-            }
-        }
-
         public TriggerClientLiveTests(bool isAsync) : base(isAsync)
         {
         }
